requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Limiter==3.9.0
Jinja2==3.1.6
python-dotenv==1.0.1
PyJWT==2.9.0
psycopg[binary]==3.2.12
gunicorn==22.0.0
google-genai==1.30.0
